Add affiliated team to post page
Teams show up hyperlinked in the sidebar, at the top for actions and at
the bottom for events. I made them different because for events, the top
of the sidebar contains the date and time, which should be up top.
For actions, the sidebar is either all of the MoC contact info or
just a button that says "Comment Now" or "Donate Now" (and either
instance can affort to be pushed down two table rows).